[Form] Fixed failing IntegerField test case and implemented true int-only support in NumberField's transformer.
Based on the failing unit test and IntegerField's own casting in its getData() method, it should truncate fractions.  Since its parent NumberField relies on NumberFormatter to power its transformer, this required implementing options for "integer_only" mode.  I have this mode override precision so not to break any compatibility with non-integer NumberFields, as well as set some attributes on the NumberFormatter to ensure that only integers are parsed (via reverseTransform() and floored values are formatted (via transform()).

An outstanding TODO was left to have reverseTranform() return a proper integer type instead of double, as it does by default.  To do this properly, we'll have to determine whether to use 32-bit or 64-bit integers, as the NumberFormatter library doesn't appear to have an option to detect (based on the architecture) this for us.  Since IntegerField::getData() casts to int already, this can be ignored for now.