[HttpFoundation] Make the class loader support multiple directories for a namespace
Here is an example to understand what this commit is about:

Zend provides Zend\Cache\Backend\Memcached.
We want to add a new cache backend called Zend\Cache\Backend\Custom.
We must keep the Zend\Cache\Backend namespace, but we can't of course
put the file in zend directory.

We put our class file in src/Application/SomeBundle/Zend/Cache/Backend/Custom.php
Our class namespace is Zend\Cache\Backend

So we want the class loader to load Zend\Cache\Backend\* from Zend
directory, plus from our application directory.

$loader->registerNamespaces(array(
    'Zend'                       => __DIR__.'/vendor/zend/library',
    'Zend\\Cache\\Backend'       => __DIR__.'/Application/SomeBundle'
));

It does not matter the order of these two lines, it won't work. Because
the autoloader stops as soon as it finds an acceptable namespace, be the
file here or not.

This commit makes the loader try all acceptable namespaces until the file
is found.
Tests included.