[Routing] Allowing array requirements to be added to a route. These are translated into a regular expression representing a "choice" between these objects.
This makes the requirements specified for parameters behave in the same way as the "magic" _method requirement. Prior to this change, the _method requirement could/should be an array, but all other requirements could not be an array.

Also, before, array requirements (other than for _method) were invalid, but would fail silently with an invalid regular expression.