[DoctrineBundle][DoctrineMongoDBBundle] Allow mappings to be loaded from vendor-organized bundles
Bundle best practices advocate organizing a bundle under a vendor namespace (e.g. "Bundle\Vendor\VendorBundle").  That puts the onus on the application developer to explicitly add "Bundle\Vendor" to the kernel's bundle directory configuration.

Alternatively, when trying to find a directory for a bundle's namespace we can exhaustively check if parent namespaces are defined in the kernel configuration, which this patch implements.