[DoctrineMongoDBBundle] Removing the $container->getParameter() calls where possible from the DI extension class.
If a parameter value is needed by the extension class itself, then $container->getParameter() call is still needed. In this case, it can be argued that those values are really DI extension "options" and should not live in the DI extension resource (they cannot be overridden via the "parameters" key like other parameters).

By removing the ->getParameter() where possible and replacing it with the %parameter.name%, the parameter value isn't processed until the container is fully built, allowing for these parameters to be overridden via the "parameters" key (which is exactly what the DIC is intended to work).