[DoctrineMongoDBBundle] Proof of concept: Migrating DI extension "options" onto an options array on the DI extension class instead of as DIC parameters.
The key to this is two-fold:

 * ->getParameter() should never be used in a DI extension class. If you're using this, you're violating how the DIC is built by consuming parameters too early.
 * Some parameters are purposefully used to perform logic in the DI extension class. In this case, these parameters aren't really DIC parameters, but are instead simply options on the DI extension class itself.

These new DI extension options are meant to be modified via the DI extension configuration (e.g. beneath doctrine_odm.mongodb). Specifying them as options on the DI extension class itself makes them very documentable and sets them apart from the DIC parameters.

As an added - potentially arguable change - you can also override any of the DI parameters by passing them beneath the DI extension, along with the DI options.

If an unrecognized DI extension option or DIC parameter is specified, a configuration exception can be thrown early to tell the user that one of his/her options is not recognized by the extension.