[DoctrineMongoDBBundle] This test is testing for the wrong behavior and forthcoming commits will fix that behavior. The actual value "mydb" should not be set here. the "default_database" is mapped to a DIC parameter and as such, the DIC parameter name should be here, not the actual end-value.
This highlights an often subtle difference between DI extension options and DIC parameters. If a config is a DI extension option, why set it as a parameter if it's not used anywhere else in the actual DI configuration? Making it a parameter allows it to be overridden naturally, but the intention really in this case is for it to be passed as an option to the DI extension. It should be ok that it's consumed immediately, because nobody else needs this value.