[DoctrineMongoDBBundle] Adding the methodology for intelligent option merging via a new "processOptionMerge" method.
This is a step closer to some of the option merging logic being centralized as the processOptionMerge() method would be the extension-specific interface for performing all custom option merging (unless I'm not forseeing some more complex situations).

The default implementation is simple: just override the previous value. The switch statement is used to illustrate how this would be expanded to be specific for any configuration options.