[Kernel] Make sure that parent bundles are registered before their descendants.
This is a bit more restrictive than the former code but I think that allowing an arbitrary order would make the code too complex while changing the order is simple.

Before this change if you register bundles in the order (GrandParent, Child, Parent), the bundleMap of GrandParent would be (Parent, Child, GrandParent).
This might hide resources from the Child Bundle overriden by both the Child and Parent Bundles.