[TwigBundle] Refactor Templating #2
Make twig environment load the template by its logical name.

Twig PHPDoc comments advertise the template name as being a string all over the place.
Then it seems better to pass the logical name of the template as a string rather than as a TemplateInterface (even thoufh Twig actually support anything as a template name).

The Loader will be in charge of converting the logical name to a template interface (which is supported without any code change).

This might result in the template logical name being parsed more often. That's why a cache has been added to the parser in order not to impact the performances.