Added PdoProfilerStorage
The SqliteProfilerStorage, which is the default storage, works fine and fast, but when receiving multiple concurrent requets, it fails due to file locking and it becomes really annoying. This is very common for AJAX powered websites.

The PdoProfilerStorage allows to store profiling information in PDO supported databases like MySQL. To enable it follow these simple steps:

1) create the database in the server, the default name is sf_profiler. It can be changed in the dsn configuration below.

2) configure PdoProfilerStorage by adding these lines to config_dev.yml:

app.config:
    profiler:
      only_exceptions: false
      storage_id: pdo
      pdo:
        dsn: mysql:host=localhost;dbname=sf_profiler
        username: root
        password: yourpassword

3) enjoy!